backend/requirements.txt
# ========================================
# KRISHOPUS-GROQ DEPENDENCIES
# ========================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Groq API
groq==0.9.0

# Document Processing
python-docx==1.1.0
python-dotenv==1.0.0

# HTTP/CORS
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
